\name{bias_function}
\alias{bias_function}
\title{Create a biased test function...}
\usage{bias_function(f, bias)
}
\description{Create a biased test function}
\details{Returns a new biased test function defined as
\deqn{g(x) = f(x) + bias.}}
\value{The biased test function.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{f}{test function}
\item{bias}{bias value.}
}

