\name{rotate_function}
\alias{rotate_function}
\title{Create a rotated test function...}
\usage{rotate_function(f, M)
}
\description{Create a rotated test function}
\details{Returns a new rotated test function defined as
\deqn{g(x) = f(Mx).}}
\value{The rotated test function.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{f}{test function.}
\item{M}{orthogonal square matrix defining the rotation.}
}

