\name{shift_function}
\alias{shift_function}
\title{shift_function}
\usage{shift_function(f, offset)
}
\description{Returns a new function
\deqn{g(x) = f(x - offset).}}
\value{The shifted test function.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{f}{test function}
\item{offset}{offset.}
}

